requirements.txt
Flask==3.1.0
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
Werkzeug==3.1.3
